+++ src/main/java/com/sdm/master/dao/RoleDAO.java
@@ -16,11 +16,11 @@ import com.sdm.master.entity.RoleEntity;
  */
 public class RoleDAO extends RestDAO {
 
-	public RoleDAO(long userId) {
+	public RoleDAO(int userId) {
 		super(RoleEntity.class.getName(), userId);
 	}
 
-	public RoleDAO(Session session, long userId) {
+	public RoleDAO(Session session, int userId) {
 		super(session, RoleEntity.class.getName(), userId);
 	}
